requirements.txt
fastapi
uvicorn
google-generativeai
google-ai-generativelanguage
python-dotenv
